+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 ### Changed
-- Do not propagate errors when invoking with useAsyncRun/useAsyncCombine*
+- Do not propagate errors when invoking with useAsyncRun/useAsyncCombine
+- The start functions in combined tasks return results
 
 ## [3.8.1] - 2019-11-28
 ### Changed